Change adminlist.txt creation to be mandatory
Consider the following example:

I add a single steamID and rebuild. I then remove the steamID and rebuild. There are now should be no steamIDs in the adminlist.txt. Although, before this change, the new state without any admins listed in the adminlist.txt would not overwrite the previous state because the option was only activated if it had 1 or more admins defined.