Add public flag option
Defaults to true when not set. This is apparently the default: 
"public 1 - Set the visibility of your server. 1 is default and will make the server visible in the browser.
Set it to 0 to make the server invisible and only joinable via the ‘Join IP’-button."
source: https://www.valheimgame.com/support/a-guide-to-dedicated-servers/